Don't kill running program if we didn't start it
Previously, the debugger would kill any program started with brickrun
even if it wasn't started from this extension. This means that the
debugger would say that console-runner was busy and not run the users
program, but also the currently running program would be killed, so it
would look like the remote device was not actually running a program.

This fixes the problem by keeping track of the debug session ID and
only sends the kill signal if the debug session actually started a
program and it has not ended yet.

+++ src/extension.ts
@@ -84,6 +84,8 @@ async function pickDevice(): Promise<void> {
     });
 }
 
+const activeDebugSessions = new Set<string>();
+
 async function handleCustomDebugEvent(event: vscode.DebugSessionCustomEvent): Promise<void> {
     let device: Device | undefined;
     switch (event.event) {
@@ -133,6 +135,7 @@ async function handleCustomDebugEvent(event: vscode.DebugSessionCustomEvent): Pr
                     else {
                         output.appendLine(`Exited with signal ${signal}.`);
                     }
+                    activeDebugSessions.delete(event.session.id);
                 });
                 channel.on('data', (chunk) => {
                     output.append(chunk.toString());
@@ -142,6 +145,7 @@ async function handleCustomDebugEvent(event: vscode.DebugSessionCustomEvent): Pr
                 });
                 output.appendLine('Started.');
                 output.appendLine('----------');
+                activeDebugSessions.add(event.session.id);
             }
             catch (err) {
                 await event.session.customRequest('ev3devBrowser.debugger.terminate');
@@ -150,7 +154,7 @@ async function handleCustomDebugEvent(event: vscode.DebugSessionCustomEvent): Pr
             break;
         case 'ev3devBrowser.debugger.stop':
             device = ev3devBrowserProvider.getDeviceSync();
-            if (device && device.isConnected) {
+            if (activeDebugSessions.has(event.session.id) && device && device.isConnected) {
                 device.exec('conrun-kill --signal=SIGKILL');
             }
             break;
